requirements.txt
azure-identity==1.23.0
azure-mgmt-support==7.0.0
azure-mgmt-resource==23.4.0
slack_sdk==3.35.0
slack_bolt==1.23.0
python-dotenv==0.21.1
cachetools==6.1.0
